fix up no-std build
